refactor: move Server to lib and flatten Game to src
