+++ src/Game.h
@@ -5,8 +5,8 @@
 
 #include "oxygine-framework.h"
 
-#include "Game/UI/Dialog.h"
-#include "src/Game/Scene/MenuScene.h"
+#include "src/UI/Dialog.h"
+#include "src/Scene/MenuScene.h"
 
 /// Height of game client, in pixels
 #define WINDOW_HEIGHT   1080
