+++ src/Spaceship/KeyboardSpaceship.h
@@ -7,8 +7,8 @@
 #include "oxygine-framework.h"
 #include "box2d.h"
 
-#include "Game/World/World.h"
-#include "Game/Spaceship/Spaceship.h"
+#include "src/World/World.h"
+#include "Spaceship.h"
 
 namespace kt {
     using namespace oxygine;
