+++ src/config.h
@@ -1,10 +1,10 @@
 #ifndef CAPSTONE_GAME_CONFIG_H
 #define CAPSTONE_GAME_CONFIG_H
 
-#include "../../config.h"
+#include "../config.h"
 
 /// Resource folder. Specifically, points to the resource XML file containing the relevant information
-#define GAME_RESOURCES  (PROJECT_ROOT "/src/Game/data/res.xml")
+#define GAME_RESOURCES  (PROJECT_ROOT "/src/data/res.xml")
 
 #endif //CAPSTONE_GAME_CONFIG_H
